Lots of minor tweaks to fix cases where reals are set to doubles, integers are set to reals etc.
git-svn-id: https://scm.physics.ox.ac.uk/svn/radtrancode/trunk@94 4dc0e162-f54d-4b9e-8a18-81c7a7f2ddce

+++ radtran/cirsrad/calc_cont.f
@@ -75,7 +75,7 @@ C******************************** CODE *********************************
 
       
       DO 13 LINE=1,NLINE(IB)
-       CURBIN = 1 + (VLIN(IB,LINE)-VBIN(1))/WING
+       CURBIN = 1 + INT((VLIN(IB,LINE)-VBIN(1))/WING)
        IGAS=IDLIN(IB,LINE)
        DO 15 J=1,NBIN
